requirements.txt
# Core ML libraries
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1

# Scientific computing
numpy==1.24.3
scipy==1.11.3

# Machine learning utilities
scikit-learn==1.3.2

# Configuration and CLI
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1

# Visualization and reporting
matplotlib==3.8.0
seaborn==0.13.0
pandas==2.1.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Voice Activity Detection
webrtcvad==2.0.10

# Optional: Advanced VAD
# pyannote.audio==3.0.1

# Utilities
h5py==3.10.0
tensorboard==2.15.1
